+++ packages/atomic/src/components/commerce/atomic-product-template/atomic-product-template.ts
@@ -7,8 +7,21 @@ import {
 } from '@coveo/headless/commerce';
 import {html, LitElement, nothing, PropertyValues} from 'lit';
 import {property, customElement, state} from 'lit/decorators.js';
-import {makeMatchConditions} from './product-template-common';
-import {ProductTemplateController} from './product-template-controller';
+import {makeMatchConditions} from '../../common/product-template/product-template-common';
+import {ProductTemplateController} from '../../common/product-template/product-template-controller';
+
+/**
+ * @alpha
+ * * A product template determines the format of the query results, depending on the conditions that are defined for each template.
+ *
+ * A `template` element must be the child of an `atomic-product-template`. Furthermore, an `atomic-commerce-product-list`, `atomic-commerce-recommendation-list`, or `atomic-commerce-search-box-instant-products` must be the parent of each `atomic-product-template`.
+ *
+ * **Note:** Any `<script>` tags that are defined inside a `<template>` element will not be executed when the products are being rendered.
+ * @MapProp name: mustMatch;attr: must-match;docs: The field and values that must be matched by a product item for the template to apply. For example, a template with the following attribute only applies to product items whose `filetype` is `lithiummessage` or `YouTubePlaylist`: `must-match-filetype="lithiummessage,YouTubePlaylist"`;type: Record<string, string[]> ;default: {}
+ * @MapProp name: mustNotMatch;attr: must-not-match;docs: The field and values that must not be matched by a product item for the template to apply. For example, a template with the following attribute only applies to product items whose `filetype` is not `lithiummessage`: `must-not-match-filetype="lithiummessage";type: Record<string, string[]> ;default: {}
+ * @slot default - The default slot where to insert the template element.
+ * @slot link - A `template` element that contains a single `atomic-product-link` component.
+ */
 
 /**
  * A product template determines the format of the query results, depending on the conditions that are defined for each template.
