+++ packages/atomic/src/components/commerce/atomic-product-template/e2e/atomic-product-template.e2e.ts
