+++ packages/atomic/src/components/commerce/atomic-product-template/e2e/fixture.ts
