+++ packages/atomic/src/components/commerce/atomic-product-template/stencil-atomic-product-template.tsx
@@ -7,7 +7,7 @@ import {MapProp} from '../../../utils/props-utils';
 import {
   ProductTemplateCommon,
   makeMatchConditions,
-} from './stencil-product-template-common';
+} from '../../common/product-template/stencil-product-template-common';
 
 /**
  * @alpha
