+++ packages/atomic/src/components/commerce/index.ts
@@ -3,3 +3,4 @@ export {AtomicCommerceInterface} from './atomic-commerce-interface/atomic-commer
 export {AtomicCommercePager} from './atomic-commerce-pager/atomic-commerce-pager.js';
 export {AtomicCommerceSearchBoxQuerySuggestions} from './atomic-commerce-search-box-query-suggestions/atomic-commerce-search-box-query-suggestions.js';
 export {AtomicCommerceSortDropdown} from './atomic-commerce-sort-dropdown/atomic-commerce-sort-dropdown.js';
+export {AtomicProductTemplate} from './atomic-product-template/atomic-product-template.js';
