+++ packages/atomic/src/components/commerce/lazy-index.ts
@@ -12,6 +12,8 @@ export default {
     await import(
       './atomic-commerce-sort-dropdown/atomic-commerce-sort-dropdown.js'
     ),
+  'atomic-product-template': async () =>
+    await import('./atomic-product-template/atomic-product-template.js'),
 } as Record<string, () => Promise<unknown>>;
 
 export type * from './index.js';
