+++ packages/atomic/src/components/commerce/product-template-components/atomic-product-field-condition/atomic-product-field-condition.tsx
@@ -4,7 +4,7 @@ import {MapProp} from '../../../../utils/props-utils';
 import {
   makeDefinedConditions,
   makeMatchConditions,
-} from '../../product-template/stencil-product-template-common';
+} from '../../../common/product-template/stencil-product-template-common';
 import {ProductContext} from '../product-template-decorators';
 
 /**
