+++ packages/atomic/src/components/common/product-template/product-template-common.spec.ts
