+++ packages/atomic/src/components/common/product-template/product-template-controller.ts
